Move tempest plugin to seperate repo
Tempest plugins rely on setuptools entrypoints and therefore
can be included in any python project as long as it properly
sets up the entrypoint. A common pattern has emerged of bundling
these plugins inside a project repo. This is likely because it
makes contributing tests at the same time as an API feature must
simpler. However, this is really an antipattern and makes consuming
the test plugins much more difficult, especially in all-in-one
deployments. In order to make using plugins more consistent and
easier for end users, packagers, and deployers this goal is to make
sure we always use a separate python project for each plugin.

Depends-On: http://gerrit.an.nuagenetworks.net/c/OpenStack/nuage-openstack-audit-tempest-plugin/+/1299
Change-Id: I31832f90cf198839d5048caf4dd12e7cfce80507